Webasyst Framework v.4.0.0   * Improved the system-wide Design section for better user experience & navigation.   * Improved support for the latest PHP versions.   * Fixed found errors.   * For developers:     * In PHP class waHtmlControl, added support for color selection fields (colorpicker).     * Added support for named placeholders for PHP function sprintf_wp().     * Added new field type color for convenient color selection in design theme settings.     * Added support for the icon_class property of design theme settings groups (group_divider)       to allow displaying custom Font Awesome icons next to settings group names in the design editor.

+++ wa-system/captcha/phpcaptcha/waPHPCaptcha.class.php
@@ -130,7 +130,7 @@ class waPHPCaptcha extends waAbstractCaptcha
         }
 
         $array_mix = preg_split('//', $str, -1, PREG_SPLIT_NO_EMPTY);
-        srand ((float)microtime()*1000000);
+        srand((int)((float)microtime()*1000000));
         shuffle ($array_mix);
         return implode("", $array_mix);
     }
